requirements.txt
streamlit
statistics
scipy.stats
numpy
LinearRegression